requirements.txt
pandas
numpy
scikit-learn
plotly
sentence-transformers
openpyxl
google-generativeai
openai